Separate out publication add/edit modal form into new component (reusable by publication.rs)
