+++ thoth-app/src/component/mod.rs
@@ -411,6 +411,7 @@ pub mod new_work;
 pub mod notification;
 pub mod prices_form;
 pub mod publication;
+pub mod publication_modal;
 pub mod publications;
 pub mod publications_form;
 pub mod publisher;
